valorisation parcours/requirements.txt
Flask
pandas
openpyxl
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
flask_sqlalchemy
